Move state.py to statemachine folder
