+++ src/isar/state_machine/states/lockdown.py
@@ -2,8 +2,8 @@ from typing import TYPE_CHECKING, List
 
 import isar.state_machine.states.home as Home
 from isar.apis.models.models import LockdownResponse
-from isar.eventhandlers.state import EventHandlerMapping, State, Transition
 from isar.models.events import EmptyMessage
+from isar.state_machine.state import EventHandlerMapping, State, Transition
 from isar.state_machine.states_enum import States
 
 if TYPE_CHECKING:
