+++ src/isar/state_machine/states/stopping_paused_return_home.py
@@ -3,8 +3,8 @@ from typing import TYPE_CHECKING, List
 import isar.state_machine.states.monitor as Monitor
 import isar.state_machine.states.return_home_paused as ReturnHomePaused
 from isar.apis.models.models import MissionStartResponse
-from isar.eventhandlers.state import EventHandlerMapping, State, Transition
 from isar.models.events import AbortedMission, EmptyMessage
+from isar.state_machine.state import EventHandlerMapping, State, Transition
 from isar.state_machine.states_enum import States
 from robot_interface.models.mission.mission import Mission
