+++ tests/isar/state_machine/states/test_await_next_mission_state.py
@@ -5,9 +5,9 @@ from typing import cast
 from pytest_mock import MockerFixture
 
 from isar.config.settings import settings
-from isar.eventhandlers.state import EventHandlerMapping, State
 from isar.modules import ApplicationContainer
 from isar.services.utilities.scheduling_utilities import SchedulingUtilities
+from isar.state_machine.state import EventHandlerMapping, State
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.await_next_mission import AwaitNextMission
 from isar.state_machine.states.paused import Paused
