+++ tests/isar/state_machine/states/test_intervention_needed_state.py
@@ -1,7 +1,7 @@
 from typing import cast
 
 from isar.config.settings import settings
-from isar.eventhandlers.state import EventHandlerMapping, State
+from isar.state_machine.state import EventHandlerMapping, State
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.going_to_lockdown import GoingToLockdown
 from isar.state_machine.states.going_to_recharging import GoingToRecharging
