+++ tests/isar/state_machine/states/test_returning_home_state.py
@@ -1,7 +1,7 @@
 from typing import cast
 
-from isar.eventhandlers.state import EventHandlerMapping, State, TimeoutHandlerMapping
 from isar.models.events import EmptyMessage
+from isar.state_machine.state import EventHandlerMapping, State, TimeoutHandlerMapping
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.await_next_mission import AwaitNextMission
 from isar.state_machine.states.home import Home
