+++ tests/isar/state_machine/states/test_stopping_paused_return_home_state.py
@@ -1,7 +1,7 @@
 from typing import cast
 
-from isar.eventhandlers.state import EventHandlerMapping, State
 from isar.models.events import EmptyMessage
+from isar.state_machine.state import EventHandlerMapping, State
 from isar.state_machine.state_machine import StateMachine
 from isar.state_machine.states.monitor import Monitor
 from isar.state_machine.states.return_home_paused import ReturnHomePaused
